+++ server/src/symbols/conditionalSymbol.ts
@@ -2,6 +2,7 @@ import { ScopedSymbol } from "antlr4-c3";
 import { IEvaluatableSymbol } from "./base";
 import { LpcTypes, SymbolKind } from "../types";
 import { CallStack, StackValue } from "../backend/CallStack";
+import { asStackValue } from "../backend/CallStackUtils";
 
 export class ConditionalSymbol
     extends ScopedSymbol
@@ -17,39 +18,14 @@ export class ConditionalSymbol
 
         if (!lhs || !rhs) return undefined;
 
-        const lhResult = lhs.eval(stack)?.value;
-        const rhResult = rhs.eval(stack)?.value;
+        const lhResult = lhs.eval(stack) as StackValue;
+        const rhResult = rhs.eval(stack) as StackValue;
 
-        switch (this.name) {
-            case "==":
-                return lhResult == rhResult;
-            case "!=":
-                return lhResult != rhResult;
-            case "<":
-                return lhResult < rhResult;
-            case ">":
-                return lhResult > rhResult;
-            case "<=":
-                return lhResult <= rhResult;
-            case ">=":
-                return lhResult >= rhResult;
-            case "|":
-                return lhResult | rhResult;
-            case "&":
-                return lhResult & rhResult;
-            case "&&":
-                return lhResult && rhResult;
-            case "||":
-                return lhResult || rhResult;
-            case "^":
-                return lhResult ^ rhResult;
-            case "in":
-                return new StackValue(1, LpcTypes.intType, this);
-            case "?":
-                return lhResult ? rhResult : undefined;
+        if (!lhResult) {
+            return undefined;
         }
 
-        throw "Conditional Symbol: operator not implemented " + this.name;
+        return lhResult.execConditional(this.name, rhResult);
     }
 
     public get kind() {