+++ server/src/symbols/expressionSymbol.ts
@@ -1,6 +1,7 @@
 import { ScopedSymbol } from "antlr4-c3";
 import { IEvaluatableSymbol, isInstanceOfIEvaluatableSymbol } from "./base";
-import { CallStack } from "../backend/CallStack";
+import { ArrayStackValue, CallStack, StackValue } from "../backend/CallStack";
+import { LpcTypes } from "../types";
 
 export class ExpressionSymbol
     extends ScopedSymbol
@@ -14,6 +15,29 @@ export class ExpressionSymbol
                 throw "not evaluable";
             }
         }
+        return scope;
+    }
+}
+
+export class BracketExpressionSymbol
+    extends ScopedSymbol
+    implements IEvaluatableSymbol<StackValue>
+{
+    eval(stack: CallStack, scope?: StackValue) {
+        let childScope: any = undefined;
+        for (const child of this.children) {
+            if (isInstanceOfIEvaluatableSymbol(child)) {
+                childScope = child.eval(stack, childScope);
+            } else {
+                throw "not evaluable";
+            }
+        }
+
+        const indexVal = childScope?.value;
+        if (indexVal != undefined && !!scope) {
+            return scope?.bracket(indexVal);
+        }
+
         return scope;
     }
 }